[BACKEND] - Criação do endpoint para buscar chamado por Id.

+++ src/main/java/com/valdir/helpdesk/repository/ChamadoRepository.java
@@ -3,5 +3,5 @@ package com.valdir.helpdesk.repository;
 import com.valdir.helpdesk.domain.Chamado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChamadoRepository extends JpaRepository<Chamado, Long> {
+public interface ChamadoRepository extends JpaRepository<Chamado, Integer> {
 }
